+++ CSSubHandle.m
@@ -82,7 +82,7 @@
 -(int)readAtMost:(int)num toBuffer:(void *)buffer
 {
 	off_t curr=[parent offsetInFile];
-	if(curr+num>end) num=end-curr;
+	if(curr+num>end) num=(int)(end-curr);
 	if(num<=0) return 0;
 	else return [parent readAtMost:num toBuffer:buffer];
 }
